requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-redis
distro==1.9.0
executing==2.0.1
requests
python-dotenv